curator_bot/requirements.txt
fastapi
uvicorn
python-dotenv
requests
watchdog
